+++ Junit/evilp0s/FileUtilTest/FileUtilTest.java
@@ -286,15 +286,5 @@ public class FileUtilTest extends SupportTest {
         FileUtil.writeAppend(new File(path + "cfile2_UTF8.TXT"), ctemp, CharsetUtil.UTF_8);
     }
 
-    @Test
-    public void testListLine() throws ParseException {
-        File file = new File(System.getProperty("user.dir") + "/Junit/Resource/time.txt");
-        List<String>  lines = FileUtil.lines(file);
-        int count=0;
-        for(String line :lines){
-            String[] tt = line.split("/");
-            count += DateUtil.subtimeBurst(tt[0], tt[1], "08:00-19:30");
-        }
-        System.out.println("===>"+count/60/60);
-    }
+
 }
